Use size_t where appropriate.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@22891 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/telnet/libtelnet/auth.c
@@ -627,7 +627,7 @@ auth_debug(int mode)
 }
 
 void
-auth_printsub(unsigned char *data, int cnt, unsigned char *buf, int buflen)
+auth_printsub(unsigned char *data, int cnt, unsigned char *buf, size_t buflen)
 {
     Authenticator *ap;
 
@@ -638,7 +638,8 @@ auth_printsub(unsigned char *data, int cnt, unsigned char *buf, int buflen)
 }
 
 void
-auth_gen_printsub(unsigned char *data, int cnt, unsigned char *buf, int buflen)
+auth_gen_printsub(unsigned char *data, int cnt,
+		  unsigned char *buf, size_t buflen)
 {
     unsigned char *cp;
     unsigned char tbuf[16];
